requirements.txt
# requirements.txt
#pip install -r requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic==2.10.6
celery==5.3.4
redis==5.0.0
alembic==1.12.0
pandas==2.1.0
psycopg2-binary==2.9.7
python-jose==3.3.0
python-multipart==0.0.6
passlib==1.7.4
python-dotenv==1.0.0